move events from write slices packages to dedicated one
