+++ src/main/java/com/dddheroes/heroesofddd/armies/write/Army.java
@@ -1,9 +1,9 @@
 package com.dddheroes.heroesofddd.armies.write;
 
 import com.dddheroes.heroesofddd.armies.write.addcreature.AddCreatureToArmy;
-import com.dddheroes.heroesofddd.armies.write.addcreature.CreatureAddedToArmy;
+import com.dddheroes.heroesofddd.armies.events.CreatureAddedToArmy;
 import com.dddheroes.heroesofddd.armies.write.addcreature.CanHaveMax7CreatureStacksInArmy;
-import com.dddheroes.heroesofddd.armies.write.removecreature.CreatureRemovedFromArmy;
+import com.dddheroes.heroesofddd.armies.events.CreatureRemovedFromArmy;
 import com.dddheroes.heroesofddd.armies.write.removecreature.CanRemoveOnlyPresentCreatures;
 import com.dddheroes.heroesofddd.armies.write.removecreature.RemoveCreatureFromArmy;
 import com.dddheroes.heroesofddd.shared.Amount;
