+++ src/main/java/com/dddheroes/heroesofddd/astrologers/write/Astrologers.java
@@ -2,7 +2,7 @@ package com.dddheroes.heroesofddd.astrologers.write;
 
 import com.dddheroes.heroesofddd.astrologers.write.proclaimweeksymbol.OnlyOneSymbolPerWeek;
 import com.dddheroes.heroesofddd.astrologers.write.proclaimweeksymbol.ProclaimWeekSymbol;
-import com.dddheroes.heroesofddd.astrologers.write.proclaimweeksymbol.WeekSymbolProclaimed;
+import com.dddheroes.heroesofddd.astrologers.events.WeekSymbolProclaimed;
 import org.axonframework.commandhandling.CommandHandler;
 import org.axonframework.eventsourcing.EventSourcingHandler;
 import org.axonframework.modelling.command.AggregateCreationPolicy;
