+++ src/main/java/com/dddheroes/heroesofddd/resourcespool/events/ResourcesWithdrawn.java
@@ -1,4 +1,4 @@
-package com.dddheroes.heroesofddd.resourcespool.write.deposit;
+package com.dddheroes.heroesofddd.resourcespool.events;
 
 import com.dddheroes.heroesofddd.resourcespool.write.ResourcesPoolEvent;
 import com.dddheroes.heroesofddd.resourcespool.write.ResourcesPoolId;
@@ -6,12 +6,12 @@ import com.dddheroes.heroesofddd.shared.Resources;
 
 import java.util.Map;
 
-public record ResourcesDeposited(
+public record ResourcesWithdrawn(
         String resourcesPoolId,
         Map<String, Integer> resources
 ) implements ResourcesPoolEvent {
 
-    public static ResourcesDeposited event(ResourcesPoolId resourcesPoolId, Resources resources) {
-        return new ResourcesDeposited(resourcesPoolId.raw(), resources.raw());
+    public static ResourcesWithdrawn event(ResourcesPoolId resourcesPoolId, Resources resources) {
+        return new ResourcesWithdrawn(resourcesPoolId.raw(), resources.raw());
     }
 }
