+++ src/test/java/com/dddheroes/heroesofddd/calendar/write/startday/StartDayTest.java
@@ -1,11 +1,12 @@
 package com.dddheroes.heroesofddd.calendar.write.startday;
 
+import com.dddheroes.heroesofddd.calendar.events.DayStarted;
 import com.dddheroes.heroesofddd.calendar.write.CalendarId;
 import com.dddheroes.heroesofddd.calendar.write.CalendarTest;
 import com.dddheroes.heroesofddd.calendar.write.Day;
 import com.dddheroes.heroesofddd.calendar.write.Month;
 import com.dddheroes.heroesofddd.calendar.write.Week;
-import com.dddheroes.heroesofddd.calendar.write.finishday.DayFinished;
+import com.dddheroes.heroesofddd.calendar.events.DayFinished;
 import com.dddheroes.heroesofddd.shared.DomainRule;
 import org.junit.jupiter.api.*;
